- Renamed FastQueue to DogArray - Renamed GrowQueue_ to DogArray_ - Renamed CircularQueue to CircularArray - DogLinkedList.previous to .prev to make it have the same number of characters as next
This was done because they are not queues but specialized arrays and have been for many years.
Dog was given as a prefix becaues GrowArray is not very original and immediately led to a conflict.
FastQueue and GrowQueue have been converging towards the same API for a while and they now officially
should have the same API. When that happens is a different story.